Selects the players we are interested in
The file contains the code to select the players we are interested in. Note that here it is used just for 2016 but can be used for any year. Make sure to include and change the path to the files otherwise it won't run. The two files required to run are in the Google Drive in the FIFA DATA folder.